requirements.txt
#python 3.11
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chumpy==0.70
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
distro==1.9.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
json-tricks==3.17.3
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mmcv==2.2.0
mmdet==3.2.0
mmengine==0.10.7
mmpose==1.3.2
mpmath==1.3.0
munkres==1.1.4
networkx==3.4.2
numpy==2.1.1
openai==1.68.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
terminaltables==3.1.10
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.94
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
xtcocotools==1.14.3
yapf==0.43.0
